Changed the server node to only control, not perform any action made a comment on how to make the navigation node. with using the information in both command server and in getandchange goal. added a comment line in getandchange_goal
Plan for the code:
command server is responsible for controlling what is happening based on a string with either start or stop.
this class keeps the track of what to do and what is done.
navigation node will make the moves and set the goals for the robot. this is resposible for starting the exploration, returning home and stoping the robot / pausing it. it listens to the  state/ to determin what to do.